+++ src/Drivers/Gd/Cloner.php
@@ -67,14 +67,14 @@ class Cloner
         // fill with background
         $processor = new ColorProcessor();
 
-        imagefill($clone, 0, 0, $processor->colorToNative($background));
+        imagefill($clone, 0, 0, $processor->export($background));
         imagealphablending($clone, true);
         imagesavealpha($clone, true);
 
         // set background image as transparent if alpha channel value if color is below .5
         // comes into effect when the end format only supports binary transparency (like GIF)
         if ($background->alpha()->value() < .5) {
-            imagecolortransparent($clone, $processor->colorToNative($background));
+            imagecolortransparent($clone, $processor->export($background));
         }
 
         return $clone;