+++ src/Drivers/Gd/Modifiers/DrawRectangleModifier.php
@@ -35,7 +35,7 @@ class DrawRectangleModifier extends GenericDrawRectangleModifier implements Spec
                     $position->y(),
                     $position->x() + $this->drawable->width(),
                     $position->y() + $this->drawable->height(),
-                    $this->driver()->colorProcessor($image)->colorToNative(
+                    $this->driver()->colorProcessor($image)->export(
                         $this->backgroundColor()
                     )
                 );
@@ -51,7 +51,7 @@ class DrawRectangleModifier extends GenericDrawRectangleModifier implements Spec
                     $position->y(),
                     $position->x() + $this->drawable->width(),
                     $position->y() + $this->drawable->height(),
-                    $this->driver()->colorProcessor($image)->colorToNative(
+                    $this->driver()->colorProcessor($image)->export(
                         $this->borderColor()
                     )
                 );