+++ src/Drivers/Imagick/Modifiers/DrawPixelModifier.php
@@ -20,7 +20,7 @@ class DrawPixelModifier extends GenericDrawPixelModifier implements SpecializedI
      */
     public function apply(ImageInterface $image): ImageInterface
     {
-        $color = $this->driver()->colorProcessor($image)->colorToNative($this->color());
+        $color = $this->driver()->colorProcessor($image)->export($this->color());
 
         try {
             $pixel = new ImagickDraw();